feat(codecompanion): conditionally register on_closed callback for ACP adapter

+++ lua/plugins/codecompanion.lua
@@ -13,18 +13,19 @@ local M = {
       group = vim.api.nvim_create_augroup("user_registering_codecompanion_callback", { clear = true }),
       callback = function(args)
         local chat = require("codecompanion").buf_get_chat(args.data.bufnr)
-        chat:add_callback("on_closed", function()
-          local nvim_pid = vim.fn.getpid()
-          local sub_pids = vim.fn.systemlist('pgrep -f "^/.*qwen.*--acp"')
-
-          for _, sub_pid in pairs(sub_pids) do
-            local ancestors = require("utils").get_ancestors(sub_pid)
-            local is_descendant = vim.tbl_contains(ancestors, nvim_pid)
-            if is_descendant then
-              vim.fn.system(("kill -9 " .. sub_pid))
+        if chat.adapter.type == "acp" and chat.adapter.formatted_name == "Qwen CLI" then
+          chat:add_callback("on_closed", function()
+            local nvim_pid = vim.fn.getpid()
+            local sub_pids = vim.fn.systemlist('pgrep -f "^/.*qwen.*--acp"')
+            for _, sub_pid in pairs(sub_pids) do
+              local ancestors = require("utils").get_ancestors(sub_pid)
+              local is_descendant = vim.tbl_contains(ancestors, nvim_pid)
+              if is_descendant then
+                vim.fn.system(("kill -9 " .. sub_pid))
+              end
             end
-          end
-        end)
+          end)
+        end
       end,
     })
     vim.api.nvim_create_autocmd("QuitPre", {
